Do not shred hard links
Before, files that were hard links were shredded
and truncated. Now, the file contents are preserved.
The link pathname is still shredded and removed.

- Fix: do not shred contents of hard links.
- Add cross-platform support for hard-link test.
- In test, expect contents are untouched for hard
  links.

+++ bleachbit/FileUtilities.py
@@ -395,7 +395,7 @@ def delete_file(path, shred):
     Returns None.
     """
     # wipe contents
-    if shred:
+    if shred and not is_hard_link(path):
         try:
             wipe_contents(path)
         except pywinerror as e:  # pylint: disable=possibly-used-before-assignment
@@ -409,6 +409,7 @@ def delete_file(path, shred):
             # permission denied (13) happens shredding MSIE 8 on Windows 7
             logger.debug("IOError #%s shredding '%s'",
                          e.errno, path)
+    if shred:
         # wipe name
         os.remove(wipe_name(path))
         return
